data_collection/requirements.txt
# Data Collection Dependencies

# HTTP Requests
requests>=2.31.0
urllib3>=2.0.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Web Scraping (for JavaScript-heavy sites)
selenium>=4.15.0

# PDF Processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0
tabula-py>=2.8.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Geocoding
geopy>=2.4.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0  # Progress bars

# Advanced Scraping Framework (optional)
scrapy>=2.11.0

# Date/Time
python-dateutil>=2.8.0

# Logging
colorlog>=6.8.0

